requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohappyeyeballs @ file:///C:/b/abs_b505trsapr/croot/aiohappyeyeballs_1725434036096/work
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
APScheduler==3.9.1.post1
asgiref==3.8.1
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_c8fgiuixkq/croot/async-timeout_1703097556097/work
attrs==22.1.0
babel==2.16.0
backports.zoneinfo==0.2.1
bidict==0.23.1
blinker @ file:///C:/b/abs_d9y2dm7cw2/croot/blinker_1696539752170/work
boto3==1.35.94
botocore==1.35.94
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
CacheControl==0.12.14
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools @ file:///C:/b/abs_792zbtc0ua/croot/cachetools_1713977157919/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi @ file:///C:/b/abs_90yq4lmu83/croot/cffi_1726856448345/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_f9ihnt72pu/croot/click_1698129847492/work
cloudpickle==2.2.1
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.8.0
cryptography==44.0.0
cycler==0.12.1
dask==2022.10.2
Django==4.2.17
djangorestframework==3.15.2
dnspython==2.3.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.7.0
flatbuffers @ file:///C:/b/abs_7f2_85svxz/croot/python-flatbuffers_1722369110965/work
fonttools==4.55.3
frozenlist @ file:///C:/b/abs_d8e__s1ys3/croot/frozenlist_1698702612014/work
fsspec==2024.12.0
future==1.0.0
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///C:/b/abs_059sl2dhu6/croot/google-auth_1715111206543/work
google-auth-oauthlib==1.0.0
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
greenlet==3.1.1
grpcio @ file:///D:/bld/grpc-split_1660349767370/work
h11==0.14.0
h5py @ file:///D:/bld/h5py_1717664890251/work
httptools==0.6.4
humanfriendly==10.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib-metadata @ file:///C:/b/abs_c1egths604/croot/importlib_metadata-suite_1704813568388/work
importlib_resources==6.4.5
jax==0.4.13
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.7
libclang==18.1.1
locket==1.0.0
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.5.3
mattermostwrapper==2.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
ml-dtypes==0.2.0
msgpack==1.1.0
multidict==5.2.0
mysql-connector-python @ file:///D:/bld/mysql-connector-python_1723222880262/work/mysql-connector-python
networkx==2.6.3
numpy==1.23.5
oauthlib @ file:///C:/b/abs_44riqpe5gg/croot/oauthlib_1679489644705/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pluggy==1.5.0
pooch @ file:///C:/b/abs_a8nmng7d_x/croot/pooch_1695850149827/work
portalocker==2.10.1
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.10
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.9
pydot==1.4.2
PyJWT @ file:///C:/b/abs_04qhgo75wz/croot/pyjwt_1715095119685/work
pykwalify==1.8.0
pymongo==4.3.3
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.4
pyreadline3==3.5.4
pyrsistent==0.20.0
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2022.7.1
pywin32==308
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
requests-oauthlib @ file:///C:/b/abs_faw0a7_wag/croot/requests-oauthlib_1720615036938/work
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.4
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
simple-websocket==1.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn-crfsuite==0.3.6
slack_sdk==3.34.0
SQLAlchemy==1.4.54
sqlparse==0.5.3
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor @ file:///C:/b/abs_16qe7jhw7n/croot/termcolor_1668084642458/work
terminaltables==3.1.10
threadpoolctl==3.5.0
toolz==1.0.0
tqdm==4.67.1
twilio==8.2.2
typing-utils==0.1.0
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.2
tzlocal==5.2
ujson @ file:///C:/b/abs_c93a278xy4/croot/ujson_1717598062082/work
urllib3==1.26.20
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
Werkzeug @ file:///C:/b/abs_8bittcw9jr/croot/werkzeug_1716533366070/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wrapt @ file:///C:/Windows/Temp/abs_7c3dd407-1390-477a-b542-fd15df6a24085_diwiza/croots/recipe/wrapt_1657814452175/work
wsproto==1.2.0
yarl @ file:///C:/b/abs_00rcuicx5k/croot/yarl_1726015884215/work
zipp @ file:///C:/b/abs_021kqswk3k/croot/zipp_1729012371206/work
